rename stream search bar to "SmartSearchBar"
