+++ src/sentry/static/sentry/app/components/smartSearchBar.jsx
